+++ integrations/meta/v1/impl.go
@@ -42,26 +42,19 @@ import (
 	"github.com/arangodb/kube-arangodb/pkg/util/svc"
 )
 
-func New(ctx context.Context, cfg Configuration) (svc.Handler, error) {
-	return newInternal(ctx, cfg)
-}
-
-func newInternal(ctx context.Context, cfg Configuration) (*implementation, error) {
-	return newInternalWithRemoteCache(ctx, cfg, cache.NewRemoteCacheWithTTL[*Object](cfg.KVCollection(cfg.Endpoint, "_system", "_meta_store"), cfg.TTL))
-}
-
-func newInternalWithRemoteCache(ctx context.Context, cfg Configuration, c cache.RemoteCache[*Object]) (*implementation, error) {
+func New(cfg Configuration) (svc.Handler, error) {
 	if err := cfg.Validate(); err != nil {
 		return nil, err
 	}
 
-	obj := &implementation{
+	return newInternal(cfg, cache.NewRemoteCacheWithTTL[*Object](cfg.KVCollection(cfg.Endpoint, "_system", "_meta_store"), cfg.TTL)), nil
+}
+
+func newInternal(cfg Configuration, c cache.RemoteCache[*Object]) *implementation {
+	return &implementation{
 		cfg:   cfg,
-		ctx:   ctx,
 		cache: c,
 	}
-
-	return obj, nil
 }
 
 var _ pbMetaV1.MetaV1Server = &implementation{}
@@ -72,7 +65,6 @@ type implementation struct {
 
 	lock sync.RWMutex
 
-	ctx context.Context
 	cfg Configuration
 
 	cache cache.RemoteCache[*Object]