+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - (Feature) (License) Activation CLI
 - (Bugfix) (DP) Propagate Timeout Across Subcommands
 - (Maintenance) Bump Dependencies
+- (Feature) (Platform) EventsV1 Integration
 
 ## [1.3.1](https://github.com/arangodb/kube-arangodb/tree/1.3.1) (2025-10-07)
 - (Documentation) Add ArangoPlatformStorage Docs & Examples